+++ apps/vmq_diversity/src/vmq_diversity_script_state.erl
@@ -179,6 +179,7 @@ handle_info({call_function, Ref, CallerPid, Function, Args}, State) ->
     CallerPid ! {call_function_response, Ref, Reply},
     {noreply, NewState}.
 
+
 %%--------------------------------------------------------------------
 %% @private
 %% @doc